a lot of fixes

+++ namenode/app.py
@@ -1,5 +1,6 @@
 from fastapi.requests import Request
 from utils.serialize.namenode import *
+from utils.serialize.server import *
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from namenode.database import *
@@ -43,21 +44,24 @@ app.add_middleware(
 @app.post('/dfs_init', response_model=InitResponse)
 async def init():
     db_init()
-    send_init_all()
     storage_list = Node.get_node_models()
+    for storage in storage_list.storages:
+        get(storage.storage_ip, 'init', Status)
     resp = InitResponse(size=storage_list.available_size())
     return resp
 
 
 @app.post('/new_node', response_model=AddNodeResponse)
 async def add_new_node(node: AddNodeRequest, request: Request):
-    if Node.q().filter(Node.storage_ip == request.client.host).count() == 0:
+    node_full_ip = f'{request.client.host}:{node.port}'
+    node_ip = request.client.host
+    if Node.q().filter(Node.storage_ip.startswith(node_ip)).count() == 0:
         storage_id = str(uuid4())
         print()
-        model = StorageModel(storage_id=storage_id, storage_ip=request.client.host, available_size=node.available_storage)
+        model = StorageModel(storage_id=storage_id, storage_ip=node_full_ip, available_size=node.available_storage)
         Node.create(model)
     else:
-        node = Node.q().filter(Node.storage_ip == request.client.host).scalar()
+        node = Node.q().filter(Node.storage_ip.startswith(node_ip)).scalar()
         storage_id = node.storage_id
         logger.info(f'Node {node.storage_id} up!')
         paths = node.up()
@@ -84,11 +88,12 @@ async def file_read(file: frmf('FileReadRequest')):
     TODO: add checking
     """
     ping_n_repl()
+    if Directory.exist(file.path):
+        raise PathIsDirectory()
     file_obj = File.query_by_paths(file.path).scalar()
     if file_obj is None:
         raise FileNotFound()
     resp = FileModel.from_orm(file_obj)
-    resp.storages = ping_all(resp.storages)
     if len(resp.storages) == 0:
         raise NodeDisconnected()
     return resp
@@ -97,6 +102,8 @@ async def file_read(file: frmf('FileReadRequest')):
 @app.post('/dfs_file_write', response_model=FileModel)
 async def file_write(file: frmf('FileWriteRequest', size=True)):
     ping_n_repl()
+    if Directory.exist(file.path):
+        raise PathIsDirectory()
     if not Directory.exist(os.path.dirname(file.path)):
         raise NoSuchDirectory()
     resp = write_file(file)
@@ -106,6 +113,8 @@ async def file_write(file: frmf('FileWriteRequest', size=True)):
 @app.post('/dfs_file_delete', response_model=FileModel)
 async def file_delete(file: frmf('FileDeleteRequest')):
     ping_n_repl()
+    if Directory.exist(file.path):
+        raise PathIsDirectory()
     if not File.exists(file.path):
         raise FileNotFound()
     resp = delete_file(file)
@@ -115,6 +124,8 @@ async def file_delete(file: frmf('FileDeleteRequest')):
 @app.post('/dfs_file_info', response_model=FileModel)
 async def file_info(file: frmf('FileInfoRequest')):
     ping_n_repl()
+    if Directory.exist(file.path):
+        raise PathIsDirectory()
     if not File.exists(file.path):
         raise FileNotFound()
     file_obj = File.query_by_paths(file.path).scalar()
@@ -124,6 +135,8 @@ async def file_info(file: frmf('FileInfoRequest')):
 @app.post('/dfs_file_copy', response_model=FileModel)
 async def file_copy(file: frmf('FileCopyRequest', new_path=True)):
     ping_n_repl()
+    if Directory.exist(file.path):
+        raise PathIsDirectory()
     if not Directory.exist(os.path.dirname(file.path)):
         raise NoSuchDirectory()
     resp = copy_file(file)
@@ -167,3 +180,17 @@ async def delete_directory(dir: DirectoryRequestModel):
     resp = get_dir_model(dir.path)
     Directory.delete(dir.path)
     return resp
+
+
+@app.post('/report', response_model=None)
+async def report(node_report: FileReport):
+    logger.info(node_report.dict())
+    if node_report.status != Status.default():
+        PendingFileNode.fail_task(node_report.path, node_report.storage_id)
+        return
+    if PendingFileNode.q().filter(
+            PendingFileNode.path == node_report.path,
+            PendingFileNode.storage_id == node_report.storage_id).count() == 0:
+        return
+    PendingFileNode.complete_task(node_report.path, node_report.storage_id)
+    return
